+++ backup_vault_notifications.tf
@@ -3,8 +3,9 @@
 
 # daily
 resource "aws_backup_vault_notifications" "daily" {
-  count             = length(var.sns_topic_arn) > 0 ? local.daily_backup_count : 0
-  backup_vault_name = aws_backup_vault.daily[0].name
+  count = var.sns_topic_arn != "" ? 1 : 0
+
+  backup_vault_name = aws_backup_vault.daily.name
   sns_topic_arn     = var.sns_topic_arn
   backup_vault_events = [
     "BACKUP_JOB_EXPIRED",